+++ frontend/.next/server/webpack-runtime.js
@@ -125,7 +125,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("91bc8e29ade212a4")
+/******/ 		__webpack_require__.h = () => ("5b22a189bbb3ea03")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/hasOwnProperty shorthand */