+++ frontend/.next/static/chunks/webpack.js
@@ -192,7 +192,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	!function() {
-/******/ 		__webpack_require__.h = function() { return "6d9e205372770389"; }
+/******/ 		__webpack_require__.h = function() { return "87d7b6660933b983"; }
 /******/ 	}();
 /******/ 	
 /******/ 	/* webpack/runtime/global */